fixed a few bugs and extracted out slug logic

+++ src/main/java/com/sixsprints/core/domain/AbstractMongoEntity.java
@@ -35,7 +35,7 @@ public class AbstractMongoEntity implements Serializable {
   protected String slug;
 
   @Indexed
-  protected Integer sequence;
+  protected Long sequence;
 
   @Indexed
   @Builder.Default
